move KaigiScaffold.kt
